% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{.boot_cep}
\alias{.boot_cep}
\title{Bootstrap consistency intervals for CEP}
\usage{
.boot_cep(y, p, n_quantiles, n_boot, prob)
}
\description{
Bootstrap consistency intervals for CEP
}
\keyword{internal}
