% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{.compute_cep}
\alias{.compute_cep}
\title{Compute CEP using isotonic regression (PAVA)}
\usage{
.compute_cep(y, p, n_quantiles)
}
\arguments{
\item{y}{Binary integer vector.}

\item{p}{Numeric vector of predicted probabilities.}

\item{n_quantiles}{Integer. Number of quantile bins.}
}
\description{
Compute CEP using isotonic regression (PAVA)
}
\keyword{internal}
