% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{.plot_calibration}
\alias{.plot_calibration}
\title{Internal plot helper for calibration}
\usage{
.plot_calibration(cep_df, boot, x_label, y_label, title, ref_line = TRUE)
}
\description{
Internal plot helper for calibration
}
\keyword{internal}
