% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ppcviz_theme}
\alias{.ppcviz_theme}
\title{Minimal ggplot2 theme consistent with bayesplot (internal)}
\usage{
.ppcviz_theme()
}
\description{
Minimal ggplot2 theme consistent with bayesplot (internal)
}
\keyword{internal}
